website: Revert accidental rename of profile_mapping page
The page for the `okta_profile_mapping` resource was renamed when the
`user_profile_mapping_source` data source was added, in
cc2e27b4463691bac4a589f043aa3794aa668e9a. Since existing links to the page and
textual references to the resource weren't updated at that time, the file rename
is the only thing that needs to be reverted.